requirements.txt
tensorflow
torch
torchvision
numpy
matplotlib
pillow
